+++ ts/packages/dispatcher/dispatcher/src/context/dispatcher/dispatcherUtils.ts
@@ -3,11 +3,11 @@
 
 import { AppAction } from "@typeagent/agent-sdk";
 import { UnknownAction } from "./schema/dispatcherActionSchema.js";
+import { DispatcherName, DispatcherEmoji } from "@typeagent/dispatcher-types";
 
-export const DispatcherName = "dispatcher";
+export { DispatcherName, DispatcherEmoji };
 export const DispatcherClarifyName = "dispatcher.clarify";
 export const DispatcherActivityName = "dispatcher.activity";
-export const DispatcherEmoji = "🤖";
 
 export function isUnknownAction(action: AppAction): action is UnknownAction {
     return action.actionName === "unknown";
