+++ ts/packages/dispatcher/types/package.json
@@ -12,7 +12,8 @@
   "author": "Microsoft",
   "type": "module",
   "exports": {
-    ".": "./dist/index.js"
+    ".": "./dist/index.js",
+    "./helpers/status": "./dist/helpers/status.js"
   },
   "files": [
     "dist",
